langgraph-threads/requirements.txt
langchain>=0.3
langchain-core>=0.3
langchain-openai>=0.3
langchain-community>=0.3
langgraph>=0.6
opentelemetry-sdk>=1.20
opentelemetry-exporter-otlp-proto-http>=1.20
opentelemetry-instrumentation-langchain>=0.36
python-dotenv>=1.0
